+++ oshi-core/src/main/java/oshi/hardware/platform/mac/MacCentralProcessor.java
@@ -30,7 +30,9 @@ import oshi.hardware.common.AbstractCentralProcessor;
 import oshi.jna.platform.mac.IOKit;
 import oshi.jna.platform.mac.SystemB;
 import oshi.jna.platform.mac.SystemB.Timeval;
+import oshi.util.ExecutingCommand;
 import oshi.util.FormatUtil;
+import oshi.util.ParseUtil;
 import oshi.util.platform.mac.IOKitUtil;
 import oshi.util.platform.mac.SysctlUtil;
 
@@ -46,6 +48,23 @@ public class MacCentralProcessor extends AbstractCentralProcessor {
     private static final long serialVersionUID = 1L;
 
     private static final Logger LOG = LoggerFactory.getLogger(MacCentralProcessor.class);
+
+    private static final long BOOTTIME;
+    static {
+        Timeval tv = new Timeval();
+        if (!SysctlUtil.sysctl("kern.boottime", tv) || tv.tv_sec == 0) {
+            // Usually this works. If it doesn't, fall back to text parsing.
+            // Boot time will be the first consecutive string of digits.
+            BOOTTIME = ParseUtil.parseLongOrDefault(
+                    ExecutingCommand.getFirstAnswer("sysctl -n kern.boottime").split(",")[0].replaceAll("\\D", ""),
+                    System.currentTimeMillis() / 1000);
+        } else {
+            // tv now points to a 64-bit timeval structure for boot time.
+            // First 4 bytes are seconds, second 4 bytes are microseconds
+            // (we ignore)
+            BOOTTIME = tv.tv_sec;
+        }
+    }
 
     /**
      * Create a Processor
@@ -156,13 +175,7 @@ public class MacCentralProcessor extends AbstractCentralProcessor {
      */
     @Override
     public long getSystemUptime() {
-        Timeval tv = new Timeval();
-        if (!SysctlUtil.sysctl("kern.boottime", tv)) {
-            return 0L;
-        }
-        // tv now points to a 16-bit timeval structure for boot time.
-        // First 8 bytes are seconds, second 8 bytes are microseconds (ignore)
-        return System.currentTimeMillis() / 1000 - tv.tv_sec;
+        return System.currentTimeMillis() / 1000 - BOOTTIME;
     }
 
     /**